Обновлена библиотека IBPP до 2.5.3.1 с исправлением заголовочного файла (добавлена библиотека cstring). Убраны преобразования типов для int32_t и все идентификаторы переведены на тип int. Проект работает в 64 битном linux.

+++ EssForms/DialogEditHronologyEA.h
@@ -26,7 +26,7 @@ public:
     HronologyEaMemModel *Model;
     QDialogButtonBox *buttonBox;
     DialogEditHronologyEA(QWidget * parent, AssociatedExtraAttribute* OneAEA);
-    long Out_value;
+    int Out_value;
 
 public slots:
 	void ExitWithAccept();
